+++ apps/avatax/src/lib/app-root-tracer.ts
@@ -1,0 +1,3 @@
+import { trace } from "@opentelemetry/api";
+
+export const appRootTracer = trace.getTracer("saleor-app-avatax");
