+++ apps/avatax/src/lib/create-instrumented-graphql-client.ts
@@ -1,6 +1,7 @@
+import { createOtelUrqlExchange } from "@saleor/apps-otel/otel-urql-exchange-factory";
 import { createGraphQLClient, CreateGraphQLClientArgs } from "@saleor/apps-shared";
 
-import { otelExchange } from "./otel-exchange";
+import { appRootTracer } from "./app-root-tracer";
 
 type CreateGraphQLClientProps = Omit<CreateGraphQLClientArgs, "opts">;
 
@@ -8,6 +9,6 @@ export const createInstrumentedGraphqlClient = (props: CreateGraphQLClientProps)
   createGraphQLClient({
     ...props,
     opts: {
-      prependingFetchExchanges: [otelExchange],
+      prependingFetchExchanges: [createOtelUrqlExchange({ tracer: appRootTracer })],
     },
   });
