+++ packages/otel/package.json
@@ -26,6 +26,7 @@
     "./batch-span-processor-factory": "./src/batch-span-processor-factory.ts",
     "./http-instrumentation-factory": "./src/http-instrumentation-factory.ts",
     "./resource-factory": "./src/resource-factory.ts",
-    "./observability-attributes": "./src/observability-attributes.ts"
+    "./observability-attributes": "./src/observability-attributes.ts",
+    "./otel-urql-exchange-factory": "./src/otel-urql-exchange-factory.ts"
   }
 }
